+++ changelog.md
@@ -1,3 +1,9 @@
+## Upcoming
+* Tweaks
+  * Added two Nausea features:
+    * Hunger when eating: if nauseous, eating will apply hunger
+    * Chance to fail to hit: if nauseous, there's a chance for an attack to miss
+
 ## 5.4.0.1
 * Added bundle to creative inventory when cloth is enabled
 * Fixed copper anvil repairs not working